refactor: 인증 의존성 정리
