+++ backend/src/main/java/com/yigongil/backend/domain/certification/CertificationRepository.java
@@ -1,5 +1,6 @@
 package com.yigongil.backend.domain.certification;
 
+import com.yigongil.backend.exception.NoCertificationException;
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
 
@@ -8,4 +9,8 @@ public interface CertificationRepository extends Repository<Certification, Long>
     Certification save(Certification feedPost);
 
     Optional<Certification> findByRoundIdAndAuthorId(Long roundId, Long memberId);
+
+    default Certification getByRoundIdAndAuthorId(Long roundId, Long memberId) {
+        return findByRoundIdAndAuthorId(roundId, memberId).orElseThrow(() -> new NoCertificationException("인증을 찾을 수 없습니다", String.valueOf(memberId)));
+    }
 }
