+++ backend/src/main/java/com/yigongil/backend/ui/doc/StudyApi.java
@@ -2,13 +2,10 @@ package com.yigongil.backend.ui.doc;
 
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.study.ProcessingStatus;
-import com.yigongil.backend.request.CertificationCreateRequest;
 import com.yigongil.backend.request.FeedPostCreateRequest;
 import com.yigongil.backend.request.StudyStartRequest;
 import com.yigongil.backend.request.StudyUpdateRequest;
-import com.yigongil.backend.response.CertificationResponse;
 import com.yigongil.backend.response.FeedPostResponse;
-import com.yigongil.backend.response.MembersCertificationResponse;
 import com.yigongil.backend.response.MyStudyResponse;
 import com.yigongil.backend.response.RoundResponse;
 import com.yigongil.backend.response.StudyDetailResponse;
@@ -202,49 +199,6 @@ public interface StudyApi {
             @Schema(hidden = true) Member member,
             @Parameter(description = "피드가 등록되는 스터디 id", required = true) Long id,
             FeedPostCreateRequest request
-    );
-
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "200"),
-                    @ApiResponse(responseCode = "400", content = @Content),
-                    @ApiResponse(responseCode = "401", content = @Content),
-                    @ApiResponse(responseCode = "404", content = @Content)
-            }
-    )
-    @SecurityRequirement(name = "token")
-    @Operation(summary = "인증 피드 등록")
-    ResponseEntity<Void> createCertification(
-            @Schema(hidden = true) Member member,
-            @Parameter(description = "피드가 등록되는 스터디 id", required = true) Long id,
-            CertificationCreateRequest request
-    );
-
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "200"),
-                    @ApiResponse(responseCode = "401", content = @Content)
-            }
-    )
-    @SecurityRequirement(name = "token")
-    @Operation(summary = "스터디 멤버 전체 인증 정보 조회")
-    ResponseEntity<MembersCertificationResponse> findAllMembersCertification(
-            @Schema(hidden = true) Member member,
-            @Parameter(description = "조회하려는 스터디 id", required = true) Long id
-    );
-
-
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "200"),
-                    @ApiResponse(responseCode = "401", content = @Content)
-            }
-    )
-    @SecurityRequirement(name = "token")
-    @Operation(summary = "스터디 멤버 단일 인증 게시글 조회")
-    ResponseEntity<CertificationResponse> findMemberCertification(
-            @Parameter(description = "인증 게시 회차 id") Long roundId,
-            @Parameter(description = "작성자 id") Long memberId
     );
 
     @ApiResponses(
