+++ backend/src/test/java/com/yigongil/backend/application/StudyServiceTest.java
@@ -30,7 +30,6 @@ class StudyServiceTest {
     private final StudyService studyService = new StudyService(
             studyRepository,
             studyMemberRepository,
-            mock(CertificationService.class),
             mock(FeedService.class),
             mock(RoundRepository.class)
     );
