+++ backend/src/test/java/com/yigongil/backend/query/profile/ui/ProfileControllerTest.java
@@ -11,8 +11,8 @@ import com.yigongil.backend.config.auth.AuthContext;
 import com.yigongil.backend.config.auth.JwtTokenProvider;
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.member.domain.MemberRepository;
-import com.yigongil.backend.domain.member.ui.MemberController;
 import com.yigongil.backend.fixture.MemberFixture;
+import com.yigongil.backend.query.profile.ProfileController;
 import com.yigongil.backend.query.profile.ProfileService;
 import com.yigongil.backend.response.ProfileResponse;
 import com.yigongil.backend.ui.exceptionhandler.InternalServerErrorMessageConverter;
